Revise the guidelines on public/private/protected
A new line above and below makes this read better. And attaching the
private/protected to the next method makes it look like it's related to
that method in particular.

Also, there's never really a need to use public directly. Instead the
following is simpler and more readable.

``` ruby
class Person
  def name
    'James'
  end

  private

  def card_number
    '4111-1111-1111-1111'
  end
end
```